Convert db initialization to event handler
